requirements.txt
PySide6>=6.6.0
websockets>=10.4
sounddevice>=0.4.6
numpy>=1.24.0
python-dotenv>=1.0.0

# OpenAI official Python SDK (for embeddings / RAG)
openai>=1.6.0
sentence-transformers

# Document parsing for company knowledge
python-docx>=0.8.11
PyPDF2>=3.0.0
pypdf
